порция говнокода

+++ 07/hw/FursAleksey/script.js
@@ -1,30 +1,50 @@
+var httpMetods = new HTTP();
+
+
 (function () {
+    var language = document.getElementById('language');
+    language = language.options[language.selectedIndex].value;
 
-    var urlSources = 'https://newsapi.org/v2/sources?language=ru&apiKey=ef4baa686f8841129d5dd8ec6cb3ca0a';
+    var category = document.getElementById('category');
+    category = category.options[category.selectedIndex].value;
 
-    var xhr = new XMLHttpRequest();
-    xhr.open('GET', urlSources);
-    xhr.send();
-
-    xhr.onload = function (){
-        var src;
-        if(xhr.readyState == 4) {
-            createButtons(JSON.parse(xhr.response).sources);
-        }
-    }
+    var urlSources = 'https://newsapi.org/v2/sources?language=' + language + '&category=' + category + '&apiKey=ef4baa686f8841129d5dd8ec6cb3ca0a';
+    httpMetods.get(urlSources, createButtons);
 }());
 
-function createButtons(src) {
+function HTTP(url, func) {
+    var xhr = new XMLHttpRequest();
+    this.get = function (url, func) {
+        xhr.open('GET', url);
+        xhr.send();
+        xhr.onload = function () {
+            if (xhr.readyState == 4) func(JSON.parse(xhr.responseText));
+        }
+    };
+    this.post = function (url, body) {
+        xhr.open('POST', url)
+        xhr.send(body);
+    };
+}
+
+function createButtons(src){
+    src = src.sources;
+
+
     var divButtons = document.getElementById('buttons');
+
     for(var i = 0; i < src.length; i++) {
         var button = document.createElement('button');
         button.innerText = src[i].name;
         button.id = src[i].id;
         addElem(button, divButtons);
     }
+
     divButtons.addEventListener('click', function (evt) {
-        deleteAllNews();
-        renderNews(evt.target.id);
+        if(evt.target.id != 'language' && evt.target.id != 'category'){
+            deleteAllNews();
+            renderNews(evt.target.id);
+        }
     });
 }
 
@@ -33,28 +53,25 @@ function addElem(elem, position) {
 }
 
 function renderNews(id) {
-    var urlHeadlines = 'https://newsapi.org/v2/top-headlines?sources=' + id + '&apiKey=ef4baa686f8841129d5dd8ec6cb3ca0a';
+    var urlHeadlines = 'https://newsapi.org/v2/top-headlines?apiKey=ef4baa686f8841129d5dd8ec6cb3ca0a&sources=' + id;
 
-    var xhr = new XMLHttpRequest();
-    xhr.open('GET', urlHeadlines);
-    xhr.send();
+    httpMetods.get(urlHeadlines, getNews);
+}
 
-    xhr.onload = function () {
-        var src = JSON.parse(xhr.response).articles;
-        console.log(src);
-        for (var i = 0; i < src.length; i++) {
-            var title = '';
-            var author = '';
-            var description = '';
+function getNews(src) {
+    src = src.articles;
+    for (var i = 0; i < src.length; i++) {
+        var title = '';
+        var author = '';
+        var description = '';
 
-            if(src[i].title) title = src[i].title;
+        if(src[i].title) title = src[i].title;
 
-            if(src[i].description) description = src[i].description;
+        if(src[i].description) description = src[i].description;
 
-            if(src[i].author) author = src[i].author;
+        if(src[i].author) author = src[i].author;
 
-            createNews(title, description, author);
-        }
+        createNews(title, description, author);
     }
 }
 
@@ -83,3 +100,4 @@ function deleteAllNews() {
     div.id = 'news';
     document.body.appendChild(div);
 }
+
